+++ cuda/UT/compassUT/include/CompassUT.cuh
@@ -25,10 +25,10 @@ __global__ void compass_ut(
   const float* dev_unique_sector_xs,
   VeloUTTracking::TrackUT* dev_compassUT_tracks,
   int* dev_atomics_compassUT,
-  int* dev_windows_layers);
+  short* dev_windows_layers);
 
 __device__ void compass_ut_tracking(
-  const int* dev_windows_layers,
+  const short* dev_windows_layers,
   uint* dev_velo_track_hits,
   const uint number_of_tracks_event,
   const int i_track,
@@ -47,12 +47,12 @@ __host__ __device__ __inline__ bool velo_track_in_UT_acceptance(
   const MiniState& state);
 
 __host__ __device__ __inline__ void fill_shared_windows(
-  const int* windows_layers,
+  const short* windows_layers,
   const uint current_track_offset,
   int* win_size_shared);
 
 __device__ __inline__ bool found_active_windows(
-  const int* dev_windows_layers,
+  const short* dev_windows_layers,
   const int total_tracks_event,
   const int track);
